rename hashing file to encryption
