+++ front/lib/api/actions/servers/salesloft/helpers.ts
@@ -21,7 +21,7 @@ import logger from "@app/logger/logger";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
-import { decrypt } from "@app/types/shared/utils/hashing";
+import { decrypt } from "@app/types/shared/utils/encryption";
 
 export async function getBearerToken(
   auth: Authenticator,
