+++ front/lib/api/assistant/conversation.ts
@@ -106,8 +106,8 @@ import type { ModelId } from "@app/types/shared/model_id";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { assertNever } from "@app/types/shared/utils/assert_never";
+import { md5 } from "@app/types/shared/utils/encryption";
 import { isString, removeNulls } from "@app/types/shared/utils/general";
-import { md5 } from "@app/types/shared/utils/hashing";
 import assert from "assert";
 import type { NextApiRequest } from "next";
 import type { Transaction } from "sequelize";
