+++ front/lib/api/dust_app_secrets.ts
@@ -1,7 +1,7 @@
 import type { Authenticator } from "@app/lib/auth";
 import { DustAppSecretModel } from "@app/lib/models/dust_app_secret";
 import type { DustAppSecretType } from "@app/types/dust_app_secret";
-import { decrypt } from "@app/types/shared/utils/hashing";
+import { decrypt } from "@app/types/shared/utils/encryption";
 import { redactString } from "@app/types/shared/utils/string_utils";
 
 export async function getDustAppSecrets(
