+++ front/lib/api/workos/user.ts
@@ -12,7 +12,7 @@ import { isDevelopment } from "@app/types/shared/env";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { isString } from "@app/types/shared/utils/general";
-import { sha256 } from "@app/types/shared/utils/hashing";
+import { sha256 } from "@app/types/shared/utils/encryption";
 import type { LightWorkspaceType } from "@app/types/user";
 import type {
   WorkOS,
