+++ front/lib/resources/user_resource.ts
@@ -17,7 +17,7 @@ import type { ModelId } from "@app/types/shared/model_id";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
-import { md5 } from "@app/types/shared/utils/hashing";
+import { md5 } from "@app/types/shared/utils/encryption";
 import type {
   LightWorkspaceType,
   UserProviderType,
