+++ front/types/shared/utils/encryption.ts
